refact: add clinPhen slice + rename types dir
