+++ src/js/features/clinPhen/makeGetIndividualData.thunk.ts
@@ -4,7 +4,7 @@ import { individualUrl } from '@/constants/configConstants';
 import type { RootState } from '@/store';
 import { printAPIError } from '@/utils/error.util';
 import { authorizedRequestConfig } from '@/utils/requests';
-import type { Individual } from '@/types/clinphen/individual';
+import type { Individual } from '@/types/clinPhen/individual';
 import { RequestStatus } from '@/types/requests';
 
 export const makeGetIndividualData = createAsyncThunk<
@@ -24,7 +24,7 @@ export const makeGetIndividualData = createAsyncThunk<
   },
   {
     condition(id, { getState }) {
-      const state = getState().query;
+      const state = getState().clinPhen;
       const cond = !state.individualDataCache[id] && state.individualDataStatus[id] !== RequestStatus.Pending;
       console.debug(
         ...(cond
