+++ src/js/types/clinPhen/gene_descriptor.ts
