+++ src/js/types/clinPhen/variant_interpretation.ts
